commit before checkout
