+++ app/dashboard/contracts/[id]/page.tsx
@@ -1,8 +1,16 @@
-const ContractDetails = () =>{
+//import Image from "next/image";
+import AddNewContract from "../new/addNewContract";
+import AllContracts from "../allContracts";
+
+
+const Home = () =>{
 
   return (
-    <h3>Coming Soon: Contract Details</h3>
+    <section className="flex text-neutral max-h-screen overflow-y-clip bg-neutral-100/50">
+      <AllContracts show={false} />
+      <AddNewContract />
+    </section>
   )
 }
 
-export default ContractDetails
+export default Home
